btc_usd_api/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.9
django-celery-beat==2.2.1
django-environ==0.8.1
django-picklefield==3.0.1
django-rest-framework==0.1.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-api-key==2.1.0
idna==3.3
kombu==5.1.0
prompt-toolkit==3.0.21
psycopg2==2.9.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
docker-compose
